add new state repository type that respects snapshops of substreams - move redis event store to its own module
