+++ src/test_helpers/redis.rs
@@ -2,7 +2,7 @@ use std::collections::HashSet;
 
 use redis_om::{redis, RedisTransportValue, StreamModel};
 
-use crate::repository::redis::{StreamModelDTO, WithSubStreamId};
+use crate::repository::redis::versioned_event::{StreamModelDTO, WithSubStreamId};
 
 use super::{
     deciders::user::{Guitar, User, UserEvent, UserId, UserName},
@@ -104,39 +104,45 @@ impl StreamModelDTO<TestUserEventDTOManager> for UserEvent {
 
     fn try_from_dto(
         model: <TestUserEventDTOManager as StreamModel>::Data,
-    ) -> Result<Self, crate::repository::redis::Error>
+    ) -> Result<Self, crate::repository::redis::versioned_event::Error>
     where
         Self: Sized,
     {
         match model.event_type {
             UserEventTypeDTO::UserAdded => match model.user {
                 Some(user) => Ok(UserEvent::UserAdded(user.into())),
-                None => Err(crate::repository::redis::Error::FromDTO(format!(
-                    "Redis UserEventDTO invalid: missing Some(User), {:?}",
-                    model
-                ))),
+                None => Err(crate::repository::redis::versioned_event::Error::FromDTO(
+                    format!(
+                        "Redis UserEventDTO invalid: missing Some(User), {:?}",
+                        model
+                    ),
+                )),
             },
             UserEventTypeDTO::UserNameUpdated => match model.user_name {
                 Some(user_name) => Ok(UserEvent::UserNameUpdated(
                     model.user_id,
                     UserName::try_from(user_name).map_err(|e| {
-                        crate::repository::redis::Error::FromDTO(format!(
+                        crate::repository::redis::versioned_event::Error::FromDTO(format!(
                             "Redis UserEventDTO invalid: {:?}",
                             e
                         ))
                     })?,
                 )),
-                None => Err(crate::repository::redis::Error::FromDTO(format!(
-                    "Redis UserEventDTO invalid: missing Some(UserName), {:?}",
-                    model
-                ))),
+                None => Err(crate::repository::redis::versioned_event::Error::FromDTO(
+                    format!(
+                        "Redis UserEventDTO invalid: missing Some(UserName), {:?}",
+                        model
+                    ),
+                )),
             },
             UserEventTypeDTO::UserGuitarAdded => match model.guitar {
                 Some(guitar) => Ok(UserEvent::UserGuitarAdded(model.user_id, guitar.into())),
-                None => Err(crate::repository::redis::Error::FromDTO(format!(
-                    "Redis UserEventDTO invalid: missing Some(Guitar), {:?}",
-                    model
-                ))),
+                None => Err(crate::repository::redis::versioned_event::Error::FromDTO(
+                    format!(
+                        "Redis UserEventDTO invalid: missing Some(Guitar), {:?}",
+                        model
+                    ),
+                )),
             },
         }
     }
